Moved the source and compiler generation to the backend
